Setting App refactoring
